fix(package): make sure import & require both work
I broke the package when I added another file: `traverse.ts`. This was because `index.ts` had an import for `./traverse.js`, but neither the CommonJS nor the ESM build had such a file because they were using `.cjs` and `.mjs` extensions. This feels like such a mess.

The best I could come up with was the use Rollup to bundle everything together and avoid the problem altogether. I also added a bunch of validation to ensure that imports and requires of the built library both work.